some edits

+++ p3.c
@@ -9,7 +9,7 @@ int main()
 	scanf("%f", &rate);
 	printf("Enter time(in years) : ");
 	scanf("%f", &time);
-	si = ((float)prin * rate * time )/ 100.0f;
+	si = (prin * rate * time )/ 100;
 	printf("Simple Interest : %.2f\n", si);
 
 }
